+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added 
+- Get groups membership by groupID BBs [#516] (https://github.com/rokwire/groups-building-block/issues/516)
+
 ## [1.53.0] - 2024-10-22
 ### Added
 - FERPA issues for group memberships [#513](https://github.com/rokwire/groups-building-block/issues/513)